+++ node_package/src/generatorFunction.ts
@@ -1,6 +1,6 @@
 // See discussion:
 // https://discuss.reactjs.org/t/how-to-determine-if-js-object-is-react-component/2825/2
-import { ComponentOrRenderFunction } from './types/index';
+import { ComponentOrRenderFunction, RenderFunction } from "./types/index";
 
 /**
  * Used to determine we'll call be calling React.createElement on the component of if this is a
@@ -10,6 +10,12 @@ import { ComponentOrRenderFunction } from './types/index';
  */
 export default function generatorFunction(component: ComponentOrRenderFunction): boolean {
   if (!component.prototype) {
+    return false;
+  }
+
+  // If zero or one args, then we know that this is a regular function that will
+  // return a React component
+  if (component.length < 2) {
     return false;
   }